server/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
python-multipart>=0.0.12
sse-starlette>=2.1
anthropic>=0.40
superdoc-sdk
python-dotenv>=1.0
